- Summary: Fixed a bug in category sorting, checked into CVS Head and Branch.(#23432)
- Bugzilla Bug (s) Resolved:

- Description:
Fixed a bug in category sorting, checked into CVS Head and Branch.
According to chart's render, if Y optional expression is specified, it means there are multiple series, each category will shows several data points. In order to get a valid chart, the chart builder was changed to limit just using category expression as category sort key. And at runtime, for sort key on category and Y optional, if the sort key equals to a expression of value series, the aggregation function of value series will be used to be set as default aggregation function of sort, otherwise sort will not have aggregation function.

- Tests Description: Manual Test

- Notes to Build Team: N/A

- Notes to Developers: N/A

- Notes to QA: N/A

- Notes to Documentation: N/A