- Summary: Fix a bug of showing DISPLAY field for dimension level.(#32714)
- Bugzilla Bug (s) Resolved:

- Description:
Fixed. The cause is the binding expression defined on report item handle is not correct. If a dimension level is defined with DISPLAY field, a special attribute 'DisplayName' should be added to binding expression, the syntax should be like dimension[DIMENSION][LEVEL]["DisplayName"], it is still using the 'DisplayName' string as the attribute name.

- Tests Description: Manual Test

- Notes to Build Team: N/A

- Notes to Developers: N/A

- Notes to QA: N/A

- Notes to Documentation: N/A