- Summary: Can not use table group as optional y grouping for chart.(#37065)
- Bugzilla Bug (s) Resolved:

- Description:
Rollbacked a previous check-in for bugzilla #312679, the Optional Y group UI shouldn't be disabled.Because for 'Inherit Columns & Groups' case, it is like 'chart sharing table' case, not all expressions of table could be used as Optional Y group of chart, but if the group number is more than two on table, the top group of table is allowed to be used as Option Y group of chart, so the Optional Y group UI shouldn't be disabled. Chart builder also provides group list for Optional Y group and context menu for data preview component to indicate which expressions could be used for Optional Y group.

- Tests Description: Manual Test

- Notes to Build Team: N/A

- Notes to Developers: N/A

- Notes to QA: N/A

- Notes to Documentation: N/A