- Bugzilla Bug(s) Resolved:
- Description:
1. Create chart.
2. Tick Flip Axis and Show X axis title.
3. Show Legend on the left.
Actual result:
Axis title is overlapped with legend entries.See attached.
Expected result:
Show legend on the right of Axis(title). 

- Resolution:
(#40606)
The cause is the algorithm for chart axis title location is wrong, now
fixed it. 
Also fixed another bug, the issue is when X axis label position is above
and X axis title is below, the axis title is cut and out of edge of
chart bottom.