Summary: Enhance data engine to support row sort on aggregation bindings
Description of Issue: In current Data Engine, we cannot do row sort on
aggregation bindings, or a binding that refers to an aggregation. 


Description of Resolution: Enhance the data engine to support sort on
aggregation bindings. Please be noted not only running aggregation, but
also
the bindings that are based on summary aggregations are supported. It
will be user's responsibility to understand the sematic meaning of such
kind of sort.[47799]