- Summary:Fix Custom tooltip does not work in live preview.(#49303)
- Bugzilla Bug(s) Resolved:
    
- Description:(#49303) There is no ActionRenderer in the RuntimeContext
when previewing the chart in the chart wizard, so the custom tooltip is
not evaluated and thus not displayed.
The fix is to refactor the tooltip evaluate code to a static function
the BIRTActionRenderer class and call it directly to evaluate the custom
tooltip when previewing in the chart wizard.