- Summary: Behaviour is incorrect using line charts with 'show line as Curve' on behalf of Shuai Zhang
- Bugzilla Bug (s) Resolved: 
 
- Description:   
reason: When there is only one single data row and rest of the data is
NULL, the chart which is show lines as curve displays no data.
solution: Keep the invalid point before the first valid point (or after
the last valid point).if it is not the first of the location point( or
the last).