- Summary:[OS][420] Doughnut chart is displayed as pie chart when preview. on behalf of Shuai Zhang
- Bugzilla Bug (s) Resolved: 
 
- Description: #50740
The reason is the size of pie chart is smaller than the inner Radius
size. if the inner radius size is bigger than pie size. when preview it
is still a pie.
Solution: set a Cap value for the inner size. let inner size always
smaller than the pie size.