- Summary: Fix issue 51307 When selecting a composite key, the selection result sometimes may not be applied to the dataset.
- Bug (s) Resolved:
51307

Description:.
Adjust the UI behaviors of refreshing parameters' positions in the data
set Parameters page.
The new rules are shown as follows:
1. The first time go the data set parameters page, we'll check the
native positions' validation. If any of the parameters has no native
position, or the native position is invalid, we'll refresh the
parameters' position as the display sequence, the same as the current
behaviors. Otherwise, we'll honor and keep their native position values.
2. If some operation has done like *Remove*, *Up* or *Down*, we'll reset
the parameters' position according to the display sequence at once.
3. When the user tries to leave the page or finish the data set editor,
the positions will not be reset once more.

- Code Reviewer: 

- Tests Description: 

- Notes to Build Team: 
N/A
- Notes to Developers: 
N/A
- Notes to QA: 
N/A
- Notes to Documentation