- Summary: Fixed a bug of legend rendering.(T51520)
- Bugzilla Bug(s) Resolved: 

- Description:
T51520
When there are many legend items, the height of chart can't display all
legend items in a column, it needs to render legend items in different
columns. But for first legend item of each column, it needs to compute
the width of legend item again, if the label of legend item is greater
than remainder width of legend, the label of legend item is needed to
reduce with ellipsis. Now fixed by using above logic.