- Summary:Fix Font Bold is always applied to chart axis title.(#51584)
- Description: For this one, the root cause is that font UI doesn't
support tri-state.

When the 'bold' is unset in the UI, the "bold" property is set to
"unset" state in the chart model.
So the hard-coded default value is used which is "bold".
That's why unsetting bold looks to have no effect.

We changing the behavior when unset the 'bold' in UI.
Other than set the value to "unset", we just set the value to "not
bold".

We can change the font UI to support tri-states in future and it will
work with the current
fix.