CheckIN: Set the default fetch size for JDBC Data Set[53409]
Description: By now we do not set the default fetch size for JDBC
drivers. So for some driver, such as Oracle (with default size 10), the
performance is not best tuned (set the size to 10000 can enhance
performance by 30 times), and for others such as MsSql/Postgress/MySql,
sometimes leads to OOM exception.

We set the default value to 10000 by now. The investigation shows that
even set it to 10 times bigger the performance benefit would be margin.
Also, compare to 1000, it still have 25% performance enhancement.