Summary: Enhance the JDBC driver to set an up limitation of fetch size.
Description: In Server environment, a high JDBC driver fetch size may
essentially leads to OOM. We explicitly set the up limit of fetch size
to 10000 in this check in. 10000 is a pretty decent number that balance
the performance and memory usage. Test in Oracle, it is 30 times faster
than default fetch size 10, but is only 1%~2% percent slower than
100000.