CheckIN:Enhance ODA-JDBC driver's handling of autocommit
Description: Different database driver have different behavior upon the
autocommit setting. For postgress, the autocommit must be set to false
in order to make the driver honor the fetch size setting. For Derby, if
autocommit is set to false then an explict call to commit must be made
prior to the connection close.

Solution: Add DB config items to handle following:
1. If the Derby driver's autocommit is set to false by user, we will do
a commit prior to the close of connection.
2. If the Postgres driver's autocommit property is not set, we will by
default set it to false to make it honor our default 10000 fetch size
setting.