Summary: Disable Drag and Drop in the Chart wizard on Mac OS and enable combo box for all data sources.(#66704)
Description:In the JFace dialog created when an item is dragged from the
Eclipse palette, DND does not work in this dialog on Mac. It could be a
SWT and JFace bug on Mac OS. So disable the DnD on Mac and enable the
combo box for all data sources.