- Summary: Error happens when create measure. 100
- Description of Issue: 
Run report with 50 dimension groups in it, it will run fine. add another
dimension group to make 51 and it will fail with "Error happens when
create measure. 100"

- Description of Resolution:
The array size was limiting the number of dimension groups. I increased
the size from 100 to 500. 

- TED(s) Resolved:
73294

- Regression ( Yes/No ):
NO

- Code Owner Team: 
Data Engine team

- Code Reviewers: 
Lin Zhu

- Project ID:

- Manual Test Description: 
Verified in local environment.

- Test Automated Cases Executed(Yes/No, if “No”, then explain why):