- Summary:
TED Issue 76305 removing a column in the SQL query results in names for other
columns being changed.

- Description of Issue:

Create a data set with SQL query that has three columns with the same name.
Example: select t1.id as "Id1", t2.id as "Id2", t3.id as "Id3" from t1, t2, t3

The column names will be automatically modified to make them unique.
The names will be id, id_1 and id_2.

Save the data set.
Create a table that uses the second and third columns (id_1 and id_2).

Now edit the SQL query and remove the first column, t1.id.
Example: select t2.id as "Id2", t3.id as "Id3" from t1, t2, t3

The column names will again be automatically modified to make them unique.
Since we removed the first column, we expect to see columns id_1 and id_2.

But, the actual new names will be id and id_1.

This means the names have changed and table bindings will be broken.

- Description of Resolution:

When the SQL query has column labels, use them as the column names.
In the example above, the column names will be unique (Id1, Id2 and Id3).
And if we remove the first column, the names will be (Id2 and Id3).

Also when obtaining the result set metadata throws an exception because the
query has a parameter, then use the prepared statement metadata.

- TED(s) Resolved: 76305
- Regression ( Yes/No ): No
- Code Owner Team: BIRT GUI
- Code Reviewers: Wei Yan
- Project ID:
- Manual Test Description: Debug in dev env
- Tests Automated Cases Executed:
- Special Notes: