Bug 489572 - Cross-site scripting (XSS) vulnerability in the __locale URL parameter of the viewer servlet.
Solution: Use ParameterAccessor.htmlEncode where appropriate in jsp page.
Also, no need to apply htmlEncode "early" when obtaining the report title.
Also, verify the locale is one of the available locales.

In general, ParameterAccessor should validate; but, not do html encoding.
The html encoding should take place when the values are used in jsp pages.

In general, this strategy has already been followed.
The exceptions are for locale/language and report title.

Signed-off-by: Carl Thronson <cthronson@actuate.com>