Fixing report item theme cut/copy/paste and drag/drop.
In a library, create two themes.
Create a report item theme in the first theme.

1.
Click the report item theme and drag it to the second theme.
Expected:
The report item theme moves from the first theme to the second theme.
Actual:
The report item theme does not move.

2.
Right-click the report item theme and choose Copy.
Right-click the second theme and choose Paste.
Expected:
A copy of the report item theme appears in the second theme.
Actual:
A copy of the report item theme appears in the Themes slot at the same
level as the main themes.

3.
Right-click the report item theme and choose Cut.
Right-click the second theme and choose Paste.
Expected:
The report item theme is no longer in the first theme and appears in the
second theme.
Actual:
The report item theme reappears in the first theme.

Signed-off-by: Carl Thronson <cthronson@actuate.com>