Nullpointer exception is thrown when we execute a report which is generated using a data design  and doesn't include default library in IHUB.
Problem:

Data set options and external meta data is not getting set for the
reports with out fincluding libraries.
This is why it is not able to find the corresponding data set and
resulting in nullpointer exception.

 Solution :

For reports which inculed library  while serializing, we were copying
the report, writing the copied report and returning the copied report
design handle. This is why all the options and externalmetadata is being
carried out. and is able to get the dataset based on these values.

The same is applied for the reports which doesn't include libraries.
In the existing code for no libraries while serializing, we were copying
the report, writing the copied report and return the original report
design handle. So changed the behaviour to return the copied report
design handle.


Signed-off-by: aswanib <aswanib@IN08282.opentext.net>