Allow multiple FactoryService implementations for a extension
Update birt.core's FactoryService extension point to allow
add an optional priority attribute. If multiple extensions
implement a FactoryService with the same id, the one with 
the highest priority will be used. This allows a factory 
service to be overridden by a patch without having to resort
to fragments.

Updated DataRequestSessionImpl class to allow it to be easily
subclassed.