Added support for Prepending and Appending PDF documents to PDF output
Adding support to append and prepend PDF documents to PDF output stream.
This was previously committed, but held up due to not being a comitter
previously.

Have the ability to append boiler plate legal documentation or financial
statements to a report, the ability to append PDF files based on a
comma-seperated list or a ArrayList defined in the Report Context has
been added.

This has been tested in actual customer environments and works. To set:
-Create a BIRT Report
-Create a User Property called AppendList.
-Set AppendList under Advanced to a Comma-Seperated list of paths to PDF
files to append

-Alternatively, create an ArrayList of String in the report context set
in global variable called appendPDF.

Signed-off-by: TheLaserNinja <36558386+TheLaserNinja@users.noreply.github.com>