Add resource filters to parent projects #699
In general, for projects that contain nested projects, filter the
folders corresponding to such nested projects such that the nested
contents of those folders are not shown in the resource tree, although
the folder itself is shown.  Of course when the nested contents are not
in the resource tree, they are not shown in any explorers.

Note that this filtering is designed not to apply for folders starting
with a '.', e.g,. .settings folders.  These folders are hidden by
default in the explorers anyway but need must be present in order to
store project-specific preferences.