UI performance degrades when selecting "many" cells in a grid/table #717
Changed to another "set selection strategy" since appending to a
existing selection yields a lot of selection changed events, which in
turn causes a lot of updates in other parts of BIRT.

Also, removed "reveal" of all the selected EditParts on the current
viewer. The logic behind this is that the user probably already sees the
EditParts which was selected. EditParts will still be revealed in other
places of BIRT.