UI performance degrades when selecting "many" cells in a grid/table #717
The "synchronization" of selections between the GEF view and the
TreeStructureViewer has previously been calling setSelection(selection,
reveal) with reveal == true. Since
AbstractTreeViewer.setSelectionOnWidget() already reveals TreeItems on
selection the extra reveal makes the tree slow and "jumpy" during
selection sync. This commit changes that behavior to only reveal the
selected items once. The down side is that if the selection is unchanged
the items will not be revealed.