UI performance degrades when selecting "many" cells in a grid/table #717 (#728)
Changed to another "set selection strategy" since appending to a
existing selection yields a lot of selection changed events, which in
turn causes a lot of updates in other parts of BIRT.

Also, removed "reveal" of all the selected EditParts on the current
viewer. The logic behind this is that the user probably already sees the
EditParts which was selected. EditParts will still be revealed in other
places of BIRT.

The "synchronization" of selections between the GEF view and the
TreeStructureViewer has previously been calling setSelection(selection,
reveal) with reveal == true. Since
AbstractTreeViewer.setSelectionOnWidget() already reveals TreeItems on
selection the extra reveal makes the tree slow and "jumpy" during
selection sync. 

This commit changes that behavior to only reveal the
selected items once. The downside is that if the selection is unchanged
the items will not be revealed.