Add support for data protocol (Question on rendering images) #657
ImageReader: Checks if an uri is using the data protocol before calling
readImage(URL) and calls readImage(byte[]) with decoded data instead.

HTML2Content: Checks if an uri is using the data protocol, if so
intercepts and sets the media type from the data uri.

Added a small DataProtocolUtil which parses url strings and returns
media type, encoding and data.