Defective PPTX format in BIRT #813 (#818)
Null guard for when DimensionType.parseUnit() fails in
PPTXUtil.convertCssToEnum(). This happens typically when the
cssDimension is "auto" for margins, which really should have a special
case treatment in the TableWriter.endCell(). However, from the look of
it, the "margin" code in endCell needs a lot of refactoring and should
perhaps be done in other places in PPTXCanvas instead ( .drawText() etc
)