Javax.xml.rpc missing from birt-runtime.zip (Was : Problem BIRT v4.9 + (#942)
Javax.xml.rpc missing from birt-runtime.zip #941

This is packaging problem in the OSGI-ification of "javax.xml.rpc".

Before this commit, javax.xml.rpc-api-1.1.2.jar was packaged,
un-exploded, inside the OSI bundle jar. That works in OSGI, via the
Export-Package and Bundle-ClassPath headers. With other Classloaders it
does not work.