Adjust getMaxQueries to return correct number (#1140)
In the case that maxStatements = 0 and maxConnections is not 0, but a finite number (e.g. 32767 in SQL Server). Then this code previously would return 1. This makes BIRT generate thousands of connections for many of our reports. This is unacceptably slow. We change the code to return 0 (=unlimited) when maxStatements=0 after this change, regardless of the number of max connections. This will in most cases have BIRT use only one connection.

In general, the code should reflect that creating Connections is expensive in most contexts.