background images based on embedded images and data-url #1175 (#1181)
1. Enhancement of background image handling

The change includes the enhancement that BIRT can handle background images based on the three source types

    URL
    Data-URL
    Report embedded images

The new option is given for:

    MasterPage
    Grid
    Table
    Row
    Cell
    TextElements
    all elements which use backgroundImage

The main changes are given on the "BackgroundImageInfo.java" and "BackgroundImageType.java".
What is important that I enhanced the constructors to but the necessary information to the background-image-object
to have all information to create the background image.

In addition to it, I made the property "backgroundImageType" accessible, which was the original design
but not fully implemented to use it on the background image side.

In addition to it, I enhanced the designer to display the images.
On the emitter side, I enhanced the HTML emitter and the PDF emitter to show the images.

Test cases of the change - done with the all-in-one-launch:
main test 1: MasterPages with all three types of background image source
main test 2: grid with grid-row-cell background images with all three types of background image source
main test 3: different text elements with all three types of background image source
main test 4: "data table with JavaScript image setting" with all three types of background image source

Attached you will find my test reports: images.zip

2. Fixing of warning/java documentation
Many of these classes had a lot of warnings based on missing java-doc, e.g., the constant-classes
and also the interface classes. I added the according comments.
I also solved some warnings based on constant usage and the render-option-hints on the HTML emitter.