Update the library versions used of dependencies used by the build
Update the setup to provide a functional targlet task that generates the
org.eclipse.birt.target.target to match.

Do not use any repository reference for the build but rather use purely
the *.target.

Use plugin-source's feature-source for feature source generation and
ensure now warnings about plugins/features without sources.

Reduce noise from Javadoc generation.

Ensure that skipTests is the default and enable those tests that
are currently running in the builds and also enable many of the other
tests that were not currently running and fix tests failing with the
latest libraries and versions.

https://github.com/eclipse/birt/issues/1286