Out of memory error when creating a chart #1727
Sometimes NumberFormat is used to handle the decimal point
and in other cases there is a dependency on ULocale.

A small fix was made to use the default ULocale in the
Numberformat calculations instead of the default NumberFormat

The whole operation is not fully understood so more issues
could arise.

See the issue for a discussion.